Fix tube map rerendering when item info dialog opens and closes
We now check to make sure that the tube map's determining props actually changed before going to re-render it.

Also, we now no longer rewrite the color schemes in place and instead properly return a new object with the new stuff under the right key.

This moves visOptions into the props of the tube map itself, because when they change the tube map needs to redraw.

We might now be able to cut some redraw calls from the setters for them in tubemap.js.