Hackily set Webpack to use Github's port forward
This adds a line to the devcontainer global bashrc that checks if the Github
Codespaces variables are set, and if so sets Webpack to use the Codespaces prot
forward for its live reload websocket.

We should maybe use a Dockerfile-based build to drop this line in the global
bashrc instead.