Use HOST and HTTPS variables to wrangle Webpack
Create React App doesn't provide a WDS_ env var to control
https://github.com/webpack/webpack-dev-server/blob/540c43852ea33f9cb18820e1cef05d5ddb86cc3e/lib/Server.js#L562-L569
when it makes the config at
https://github.com/facebook/create-react-app/blob/20edab4894b301f6b90dad0f90a2f82c52a7ac66/packages/react-scripts/config/webpackDevServer.config.js#L19-L22

So we have to actually put the server into HTTPS mode if we want WSS
used for the hot reload server. Luckily, Github's proxy ignores
self-signed certs. But we still need to set HOST or the proxy weirdly
doesn't work with HTTPS ports. See
https://github.com/orgs/community/discussions/28563#discussioncomment-4984423